feat(brdb,brz): refactor pendingfs to support brdb and brz conversions
