+++ crates/brdb/examples/brdb_to_brz.rs
@@ -1,0 +1,13 @@
+use brdb::{Brdb, Brz};
+use std::path::PathBuf;
+
+/// Reads a brdb and outputs a brz
+fn main() -> Result<(), Box<dyn std::error::Error>> {
+    let src = PathBuf::from("./prefab.brdb");
+    let dst = PathBuf::from("./prefab.brz");
+
+    let pending = Brdb::open(src)?.to_pending()?;
+    Brz::write_pending(dst, pending)?;
+
+    Ok(())
+}
